change_user_id_type_in_database

+++ tables.sql
@@ -1,7 +1,7 @@
 -- users_info
 CREATE TABLE users_info (
     id SERIAL PRIMARY KEY,
-    user_id INTEGER UNIQUE, -- или PRIMARY KEY,
+    user_id BIGINT UNIQUE, -- или PRIMARY KEY,
     age INTEGER,
     sex TEXT,
     lang CHARACTER VARYING
@@ -10,7 +10,7 @@ CREATE TABLE users_info (
 -- user_moods
 CREATE TABLE user_moods (
     id SERIAL PRIMARY KEY,
-    user_id INTEGER,
+    user_id BIGINT,
     mood INTEGER,
     date DATE,
     time TIME WITHOUT TIME ZONE,
